changed data model
