+++ src/NHSDPublicBrowseAcceptanceTests.TestData/Solutions/HostingTypeStrings.cs
@@ -3,6 +3,6 @@
     internal static class HostingTypeStrings
     {
         internal const string CompleteHostingTypes =
-            "{\"PublicCloud\":{\"Summary\":\"dasfdasf\",\"Link\":\"http://www.bbc.co.uk\",\"RequiresHSCN\":\"This Solution requires a HSCN/N3 connection\"},\"PrivateCloud\":{\"Summary\":\"asdfasdfdasf\",\"Link\":\"http://www.bbc.co.uk\",\"HostingModel\":\"adfadfasd\",\"RequiresHSCN\":\"This Solution requires a HSCN/N3 connection\"},\"HybridHostingType\":{\"Summary\":\"asdfasdfasdf\",\"Link\":\"https://www.bbc.co.uk\",\"HostingModel\":\"asdfasdfasdf\",\"RequiresHSCN\":\"This Solution requires a HSCN/N3 connection\"},\"OnPremise\":{\"Summary\":\"sdfasdfasdfasdf\",\"Link\":\"https://www.bbc.co.uk\",\"HostingModel\":\"sdfasdfasdf\",\"RequiresHSCN\":\"This Solution requires a HSCN/N3 connection\"}}";
+            "{\"hosting\":{\"publicCloud\":{\"summary\":\"dasfdasf\",\"urlLink\":\"http://www.bbc.co.uk\",\"hscnN3AccessRequired\":\"Link to HSCN or N3 network required to access service\"},\"PrivateCloud\":{\"summary\":\"asdfasdfdasf\",\"urlLink\":\"http://www.bbc.co.uk\",\"hostingOverview\":\"adfadfasd\",\"hscnN3AccessRequired\":\"Link to HSCN or N3 network required to access service\"},\"hybrid\":{\"summary\":\"asdfasdfasdf\",\"urlLink\":\"https://www.bbc.co.uk\",\"hostingOverview\":\"asdfasdfasdf\",\"hscnN3AccessRequired\":\"Link to HSCN or N3 network required to access service\"},\"onPremise\":{\"summary\":\"sdfasdfasdfasdf\",\"urlLink\":\"https://www.bbc.co.uk\",\"hostingOverview\":\"sdfasdfasdf\",\"hscnN3AccessRequired\":\"Link to HSCN or N3 network required to access service\"}}}";
     }
 }
