--- src/NHSDPublicBrowseAcceptanceTests.TestData/Solutions/SolutionDetail.cs
@@ -1,47 +1,0 @@
-﻿using System;
-using System.Linq;
-using NHSDPublicBrowseAcceptanceTests.TestData.Utils;
-
-namespace NHSDPublicBrowseAcceptanceTests.TestData.Solutions
-{
-    public sealed class SolutionDetail
-    {
-        public Guid SolutionDetailId { get; set; }
-        public string SolutionId { get; set; }
-        public string AboutUrl { get; set; }
-        public string Features { get; set; }
-        public string ClientApplication { get; set; }
-        public string Summary { get; set; }
-        public string FullDescription { get; set; }
-        public string RoadMap { get; set; }
-        public string HostingTypes { get; set; }
-        public string IntegrationsUrl { get; set; }
-        public string ImplementationTimescales { get; set; }
-        public Guid LastUpdatedBy { get; set; } = Guid.Empty;
-        public DateTime LastUpdated { get; set; } = DateTime.Now;
-
-        public SolutionDetail Retrieve(string connectionString)
-        {
-            var query = Queries.GetSolutionDetail;
-            return SqlExecutor.Execute<SolutionDetail>(connectionString, query, this).Single();
-        }
-
-        public void Create(string connectionString)
-        {
-            var query = Queries.CreateSolutionDetail;
-            SqlExecutor.Execute<SolutionDetail>(connectionString, query, this);
-        }
-
-        public void Update(string connectionString)
-        {
-            var query = Queries.UpdateSolutionDetail;
-            SqlExecutor.Execute<SolutionDetail>(connectionString, query, this);
-        }
-
-        public void Delete(string connectionString)
-        {
-            var query = Queries.DeleteSolutionDetail;
-            SqlExecutor.Execute<SolutionDetail>(connectionString, query, this);
-        }
-    }
-}
