+++ src/NHSDPublicBrowseAcceptanceTests.Tests/Steps/BrowseSolutions/ViewASolution.cs
@@ -41,11 +41,9 @@ namespace NHSDPublicBrowseAcceptanceTests.Tests.Steps.BrowseSolutions
             new ViewSolutionsList(_test, _context).GivenThatAUserHasChosenToViewAListOfAllSolutions();
             var oldDate = new DateTime(2001, 02, 03);
             LastUpdatedHelper.UpdateLastUpdated(oldDate, "Solution", "id", _test.Solution.Id, _test.ConnectionString);
-            LastUpdatedHelper.UpdateLastUpdated(oldDate, "SolutionDetail", "SolutionId", _test.Solution.Id,
-                _test.ConnectionString);
             LastUpdatedHelper.UpdateLastUpdated(oldDate, "MarketingContact", "SolutionId", _test.Solution.Id,
                 _test.ConnectionString);
-            _test.Pages.SolutionsList.OpenNamedSolution(_test.Solution.Name);
+            _test.Pages.SolutionsList.OpenNamedSolution(_test.CatalogueItem.Name);
         }
 
 
@@ -63,7 +61,7 @@ namespace NHSDPublicBrowseAcceptanceTests.Tests.Steps.BrowseSolutions
             _test.Pages.ViewASolution.PageDisplayed(_test.Url);
             var id = _test.Pages.ViewASolution.GetSolutionId();
             _test.Solution = new Solution {Id = id}.Retrieve(_test.ConnectionString);
-            _test.SolutionDetail = new SolutionDetail {SolutionId = _test.Solution.Id}.Retrieve(_test.ConnectionString);
+            _test.CatalogueItem = new CatalogueItem { CatalogueItemId = id }.Retrieve(_test.ConnectionString);
         }
 
         [Then(@"the page will contain Supplier Name")]
@@ -76,13 +74,13 @@ namespace NHSDPublicBrowseAcceptanceTests.Tests.Steps.BrowseSolutions
         public void ThenSolutionName()
         {
             var solutionName = _test.Pages.ViewASolution.GetSolutionName();
-            solutionName.Should().Be(_test.Solution.Name);
+            solutionName.Should().Be(_test.CatalogueItem.Name);
         }
 
         [Then(@"Solution Summary")]
         public void ThenSolutionSummary()
         {
-            var expected = Regex.Replace(_test.SolutionDetail.Summary, @"\s", "");
+            var expected = Regex.Replace(_test.Solution.Summary, @"\s", "");
             var solutionSummary = Regex.Replace(_test.Pages.ViewASolution.GetSolutionSummary(), @"\s", "");
             solutionSummary.Should().Be(expected);
         }
@@ -90,7 +88,7 @@ namespace NHSDPublicBrowseAcceptanceTests.Tests.Steps.BrowseSolutions
         [Then(@"Solution Full Description")]
         public void ThenSolutionFullDescription()
         {
-            var expected = Regex.Replace(_test.SolutionDetail.FullDescription, @"\s", "");
+            var expected = Regex.Replace(_test.Solution.FullDescription, @"\s", "");
             var solutionFullDescription = Regex.Replace(_test.Pages.ViewASolution.GetSolutionFullDescription(), @"\s", "");
             solutionFullDescription.Should().Be(expected);
         }
@@ -99,10 +97,10 @@ namespace NHSDPublicBrowseAcceptanceTests.Tests.Steps.BrowseSolutions
         [Then(@"About Solution URL")]
         public void ThenAboutSolutionURL()
         {
-            if (!string.IsNullOrEmpty(_test.SolutionDetail.AboutUrl))
+            if (!string.IsNullOrEmpty(_test.Solution.AboutUrl))
             {
                 var solutionAboutUrl = _test.Pages.ViewASolution.GetSolutionAboutUrl();
-                solutionAboutUrl.Should().Be(_test.SolutionDetail.AboutUrl);
+                solutionAboutUrl.Should().Be(_test.Solution.AboutUrl);
             }
         }
